♻️ [Refactor] Diary 이미지 뷰모델로 분리
